+++ built/jtree.js
@@ -18,6 +18,5 @@ jtree.TerminalNode = GrammarBackedTerminalNode_1.default;
 jtree.AnyNode = GrammarBackedAnyNode_1.default;
 jtree.GrammarProgram = GrammarProgram_1.default;
 jtree.TreeNotationCodeMirrorMode = TreeNotationCodeMirrorMode_1.default;
-jtree.getLanguage = name => require(__dirname + `/../langs/${name}/index.js`);
 jtree.getVersion = () => "20.0.0";
 exports.default = jtree;