+++ src/jtree.ts
@@ -17,7 +17,6 @@ class jtree {
   static AnyNode = GrammarBackedAnyNode
   static GrammarProgram = GrammarProgram
   static TreeNotationCodeMirrorMode = TreeNotationCodeMirrorMode
-  static getLanguage = name => require(__dirname + `/../langs/${name}/index.js`)
   static getVersion = () => "20.0.0"
 }